support Module.to and change the alias for paddle.functional.split
